DataDownloader/requirements.txt
py5paisa
python-dotenv
pyotp